[#164][reservation] refactor: dto 패키지 이동
