+++ reservation/src/main/java/table/eat/now/reservation/reservation/application/service/ReservationService.java
@@ -5,10 +5,10 @@
 package table.eat.now.reservation.reservation.application.service;
 
 import table.eat.now.reservation.reservation.application.service.dto.request.CancelReservationCommand;
+import table.eat.now.reservation.reservation.application.service.dto.request.ConfirmReservationCommand;
 import table.eat.now.reservation.reservation.application.service.dto.request.CreateReservationCommand;
 import table.eat.now.reservation.reservation.application.service.dto.request.GetReservationCriteria;
 import table.eat.now.reservation.reservation.application.service.dto.response.CancelReservationInfo;
-import table.eat.now.reservation.reservation.application.service.dto.response.ConfirmReservationCommand;
 import table.eat.now.reservation.reservation.application.service.dto.response.CreateReservationInfo;
 import table.eat.now.reservation.reservation.application.service.dto.response.GetReservationInfo;
