+++ reservation/src/main/java/table/eat/now/reservation/reservation/application/service/dto/request/ConfirmReservationCommand.java
@@ -1,4 +1,4 @@
-package table.eat.now.reservation.reservation.application.service.dto.response;
+package table.eat.now.reservation.reservation.application.service.dto.request;
 
 import lombok.Builder;
